Relocates data models to respective modules
Moves `Identity` and `ExperienceBuilder` data models from shared module and pipeline to the modules where they are utilized.
This improves code organization and encapsulation by placing data structures closer to their usage context.
