+++ tests/test_pipeline_highlevel_coverage.py
@@ -4,108 +4,7 @@ import pytest
 import json
 from pathlib import Path
 from unittest.mock import Mock, patch, MagicMock
-from cvextract.pipeline_highlevel import Identity, ExperienceBuilder, extract_cv_structure, process_single_docx
-
-
-class TestIdentity:
-    """Tests for Identity dataclass."""
-
-    def test_identity_as_dict(self):
-        """Test Identity.as_dict() method conversion."""
-        identity = Identity(
-            title="Software Engineer",
-            full_name="John Doe",
-            first_name="John",
-            last_name="Doe"
-        )
-        result = identity.as_dict()
-        
-        assert result == {
-            "title": "Software Engineer",
-            "full_name": "John Doe",
-            "first_name": "John",
-            "last_name": "Doe",
-        }
-
-    def test_identity_frozen(self):
-        """Test that Identity is immutable (frozen)."""
-        identity = Identity(
-            title="Engineer",
-            full_name="Jane Smith",
-            first_name="Jane",
-            last_name="Smith"
-        )
-        
-        with pytest.raises(AttributeError):
-            identity.title = "Updated"
-
-
-class TestExperienceBuilder:
-    """Tests for ExperienceBuilder dataclass."""
-
-    def test_experience_builder_finalize_basic(self):
-        """Test finalize() converts builder to dict with basic data."""
-        builder = ExperienceBuilder(
-            heading="Jan 2020 - Present",
-            description_parts=["Developed features"],
-            bullets=["Built APIs", "Fixed bugs"],
-            environment=["Python", "FastAPI"]
-        )
-        
-        result = builder.finalize()
-        
-        assert result == {
-            "heading": "Jan 2020 - Present",
-            "description": "Developed features",
-            "bullets": ["Built APIs", "Fixed bugs"],
-            "environment": ["Python", "FastAPI"],
-        }
-
-    def test_experience_builder_finalize_empty_environment(self):
-        """Test finalize() sets environment to None when empty."""
-        builder = ExperienceBuilder(
-            heading="  Senior Engineer  ",
-            description_parts=["Managed team", "Led projects"],
-            bullets=["Mentored"],
-            environment=[]
-        )
-        
-        result = builder.finalize()
-        
-        assert result["heading"] == "Senior Engineer"
-        assert result["description"] == "Managed team Led projects"
-        assert result["environment"] is None
-
-    def test_experience_builder_finalize_whitespace_handling(self):
-        """Test finalize() strips heading and final joined description."""
-        builder = ExperienceBuilder(
-            heading="  2019 - 2020  ",
-            description_parts=["  Part A  ", "  Part B  "],
-            bullets=["  • Item  "],
-            environment=["C++"]
-        )
-        
-        result = builder.finalize()
-        
-        assert result["heading"] == "2019 - 2020"
-        # Join preserves internal spacing, then strip removes outer
-        assert result["description"] == "Part A     Part B"
-        assert result["bullets"] == ["  • Item  "]  # bullets not stripped
-
-    def test_experience_builder_multiple_bullets(self):
-        """Test finalize() preserves multiple bullet points."""
-        bullets = ["First task", "Second task", "Third task"]
-        builder = ExperienceBuilder(
-            heading="Job",
-            description_parts=["desc"],
-            bullets=bullets,
-            environment=["Python"]
-        )
-        
-        result = builder.finalize()
-        
-        assert result["bullets"] == bullets
-        assert len(result["bullets"]) == 3
+from cvextract.pipeline_highlevel import extract_cv_structure, process_single_docx
 
 
 class TestExtractCvStructure:
