+++ tests/test_shared.py
@@ -39,29 +39,3 @@ class TestXMLSanitization:
         assert out["b"][1]["c"] == "oknope"
         assert out["d"] == 123
 
-
-class TestIdentity:
-    """Tests for Identity dataclass."""
-    
-    def test_identity_as_dict_returns_all_fields(self):
-        """Identity object should convert to dictionary with all fields."""
-        ident = cvextract.shared.Identity(title="Senior", full_name="Ada Lovelace", first_name="Ada", last_name="Lovelace")
-        assert ident.as_dict() == {
-            "title": "Senior",
-            "full_name": "Ada Lovelace",
-            "first_name": "Ada",
-            "last_name": "Lovelace",
-        }
-
-
-class TestExperienceBuilder:
-    """Tests for ExperienceBuilder."""
-    
-    def test_finalize_with_no_environment_returns_none_for_environment(self):
-        """When environment list is empty, finalized dict should have environment=None."""
-        b = cvextract.shared.ExperienceBuilder(heading="H", description_parts=["a", "b"], bullets=["x"])
-        d = b.finalize()
-        assert d["heading"] == "H"
-        assert d["description"] == "a b"
-        assert d["bullets"] == ["x"]
-        assert d["environment"] is None
